+++ covalent/_dispatcher_plugins/local.py
@@ -20,7 +20,7 @@
 
 from copy import deepcopy
 from functools import wraps
-from typing import Callable, Dict
+from typing import Callable, Dict, List, Optional
 
 import requests
 
@@ -30,7 +30,37 @@ from .._results_manager.results_manager import get_result
 from .._shared_files.config import get_config
 from .._workflow.lattice import Lattice
 from .base import BaseDispatcher
-from .redispatch_utils import get_request_body
+
+
+def get_request_body(
+    dispatch_id: str,
+    new_args: Optional[List] = None,
+    new_kwargs: Optional[Dict] = None,
+    replace_electrons: Optional[Dict[str, Callable]] = None,
+    reuse_previous_results: bool = False,
+) -> Dict:
+    """Get request body for re-dispatching a workflow."""
+    if new_args is None:
+        new_args = []
+    if new_kwargs is None:
+        new_kwargs = {}
+    if replace_electrons is None:
+        replace_electrons = {}
+    if new_args or new_kwargs:
+        res = get_result(dispatch_id)
+        lat = res.lattice
+        lat.build_graph(*new_args, **new_kwargs)
+        json_lattice = lat.serialize_to_json()
+    else:
+        json_lattice = None
+    updates = {k: v.electron_object.as_transportable_dict for k, v in replace_electrons.items()}
+
+    return {
+        "json_lattice": json_lattice,
+        "dispatch_id": dispatch_id,
+        "electron_updates": updates,
+        "reuse_previous_results": reuse_previous_results,
+    }
 
 
 class LocalDispatcher(BaseDispatcher):
